third_party/WebKit/LayoutTests/fast/dom/shadow/dynamically-added-deep-combinator-and-shadow-pseudo-element-expected.txt

CONSOLE WARNING: line 311: /deep/ combinator is deprecated and will be a no-op in M60, around August 2017. See https://www.chromestatus.com/features/4964279606312960 for more details.
CONSOLE WARNING: line 311: ::shadow pseudo-element is deprecated and will be removed in M60, around August 2017. See https://www.chromestatus.com/features/6750456638341120 for more details.
Should have blue background.
PASS backgroundColorOf('host/target1') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host/target2') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host/target3') is "rgb(0, 0, 255)"
PASS backgroundColorOf('host/target4') is "rgb(0, 0, 255)"
PASS backgroundColorOf('target5') is "rgb(0, 0, 255)"
PASS successfullyParsed is true

TEST COMPLETE
